コアシステムの修正
プレイヤーにバフデバフを与えていたところをボスに与えるように変更
・ 攻撃力、攻撃速度、被ダメージ倍率をいじるように変更しました
・ エフェクトとアニメーションの速度をいじれるようにした
・ 実際に値の変化等起きることは確認済み
・ UIについてはまだ追加していません

+++ GameTemplate/Game/src/Actor/BossCharacter.cpp
@@ -28,10 +28,10 @@ void BossCharacter::SetupAnimation()
 	}
 }
 
-void BossCharacter::PlayAnimation(const int id,const int animSpeed)
+void BossCharacter::PlayAnimation(const int id,const float animSpeed)
 {
 	modelRender_.PlayAnimation(id); // Idをもとにそのアニメーションの再生
-	modelRender_.SetAnimationSpeed(animSpeed);
+	modelRender_.SetAnimationSpeed(animSpeed); // 小数(0.5倍速など)も正しく反映されるようfloatで受け取る
 }
 
 /** ===================================================== */
@@ -136,7 +136,7 @@ bool BossCharacter::Start()
 void BossCharacter::Update()
 {
 	// アクティブでないなら更新しない
-	if (!isUpdate_) return; 
+	if (!isUpdate_) return;
 
 
 	// ボスの移動処理
