+++ GameTemplate/Game/src/Actor/BossCharacter.h
@@ -39,7 +39,7 @@ public:
 	/** アニメーションの読み込み */
 	void SetupAnimation();
 	/** アニメーションの再生 */
-	void PlayAnimation(const int id, const int animSpeed = 1.0f);
+	void PlayAnimation(const int id, const float animSpeed = 1.0f);
 
 
 /** =================================================================== */
